code refactor on models
