+++ lib/features/home/data/model/widgets_model/currency_exchange_parameters_model.dart
