+++ lib/features/home/data/remote_data_source/currency_exchange_data_source.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 
 import '../../../../core/api/dio_consumer.dart';
 import '../../../../core/api/end_points.dart';
-import '../model/currency_exchange_parameters_model.dart';
+import '../model/widgets_model/currency_exchange_parameters_model.dart';
 
 abstract class CurrencyExchangeDataSource {
   Future<Response<dynamic>> getCurrencyExchange({
