+++ lib/features/home/data/repository/currency_exchange_repository.dart
@@ -6,7 +6,7 @@ import '../../../../core/error/error_handler.dart';
 import '../../../../core/utils/service_locator.dart';
 import '../model/currency_exchange_models/all_currencies_model.dart';
 import '../model/currency_exchange_models/fluctuation_currencies_model.dart';
-import '../model/currency_exchange_parameters_model.dart';
+import '../model/widgets_model/currency_exchange_parameters_model.dart';
 import '../remote_data_source/currency_exchange_data_source.dart';
 
 abstract class CurrencyExchangeRepository {
